WIP Fix intersecting cycles detection
A node being part of a cycle does not mean it cannot be part of another cycle,
so stopping the search if node.name in self.cyclenodes stop the search too soon
when multiple cycle sharea node.

An easy case is a triangular import

a.py
- from . import b, c
b.py
- from . import a, b
c.py
- from . import a, b

Using --show-cycle on this showed only one cycle.